% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{cross_validate}
\alias{cross_validate}
\title{this function takes the train structure and hyperparameters, and an integer n
and does an n fold cross validation and returns the result.}
\usage{
cross_validate(train_structure, hyperparameters, nfold = 5)
}
\arguments{
\item{train_structure}{the data structure generated by prepare_training_set}

\item{hyperparameters}{the list generated by guess_hyperparameters function}

\item{nfold}{number of folds of CV}
}
\description{
this function takes the train structure and hyperparameters, and an integer n
and does an n fold cross validation and returns the result.
}
