% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_data.R
\name{get_class}
\alias{get_class}
\title{this function helps compute the class of a given column
and returns only one character value as a result.}
\usage{
get_class(column)
}
\arguments{
\item{column}{of the data frame whose class we are interested in.}
}
\description{
this function helps compute the class of a given column
and returns only one character value as a result.
}
