% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_data.R
\name{get_normalizing_factors}
\alias{get_normalizing_factors}
\title{this function returns the factors by which a df was normalized,
so that the test dataset can be processed similarly}
\usage{
get_normalizing_factors(df, target_variable = NA)
}
\arguments{
\item{df}{a data frame}

\item{target_variable}{a string name for the target variable}
}
\value{
data frame with the normalizing factors
}
\description{
this function returns the factors by which a df was normalized,
so that the test dataset can be processed similarly
}
