% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{get_predictions}
\alias{get_predictions}
\title{this function takes the model structure generated by train_model, along with
a test set in the same format as the untransformed *input* df
to the prepare_training_set function, to return a prediction vector
in the untransformed df.}
\usage{
get_predictions(model_structure, test_df)
}
\description{
this function takes the model structure generated by train_model, along with
a test set in the same format as the untransformed *input* df
to the prepare_training_set function, to return a prediction vector
in the untransformed df.
}
