% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{get_predictions_linear}
\alias{get_predictions_linear}
\title{this function takes the model structure generated by train_linear_model, along with
a test set in the same format as the untransformed *input* df
to the prepare_training_set function, to return a prediction vector
in the untransformed df.}
\usage{
get_predictions_linear(model_structure, test_df)
}
\arguments{
\item{model_structure}{model structure cresated earlier}

\item{test_df}{the test data frame}
}
\description{
this function takes the model structure generated by train_linear_model, along with
a test set in the same format as the untransformed *input* df
to the prepare_training_set function, to return a prediction vector
in the untransformed df.
}
