% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_data.R
\name{get_train_levels}
\alias{get_train_levels}
\title{This function returns a data frame of all the various categories in the
training set, so that these are available while pre-processing the test set}
\usage{
get_train_levels(df)
}
\arguments{
\item{df}{the training set data frame}
}
\value{
a data frame with all the columns and all levels in the categorical columns
}
\description{
This function returns a data frame of all the various categories in the
training set, so that these are available while pre-processing the test set
}
