% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{guess_hyperparameters}
\alias{guess_hyperparameters}
\title{This function guesses sensible parameters for setting up the
xgb model. If the user has supplied parameters, the function
chooses the ones least likely to lead to overfitting.}
\usage{
guess_hyperparameters(train_structure, n_estimators = 10,
  learning_rate = 0.05, depth = 2, nrounds = 100,
  objective_function = NA, eval_metric = NA)
}
\arguments{
\item{train_structure}{is the data structure produced by prepare_training_set}

\item{n_estimators}{minimum number of estimators}

\item{learning_rate}{maximum learning rate}

\item{depth}{minimum depth of trees}

\item{nrounds}{the number of rounds of training}

\item{objective_function}{the objective function to be used}

\item{eval_metric}{the evaluation metric}
}
\description{
Objective functions and eval metrics are always taken
to be the user supplied values, if the suer has supplied them,
else they are guessed based on the type of the target variable.
}
