% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_data.R
\name{handle_missing_values}
\alias{handle_missing_values}
\title{tThis function implements a simple NA handling system for
categorical and numeric variables.}
\usage{
handle_missing_values(df, target_variable = "y", train_facs)
}
\arguments{
\item{df}{a data frame}

\item{target_variable}{a string name for the target variable}
}
\value{
list of data frames
}
\description{
tThis function implements a simple NA handling system for
categorical and numeric variables.
}
