% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{handle_new_test_levels}
\alias{handle_new_test_levels}
\title{this functions deals with categorical variables in the test set
which are not present in the training set.}
\usage{
handle_new_test_levels(model_structure, features)
}
\arguments{
\item{model_structure}{the model strucxture created by xgb}

\item{features}{is the output of xgboost::xgb.DMatrix for the test set after combining with the levels from training set}
}
\description{
this functions deals with categorical variables in the test set
which are not present in the training set.
}
