% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_data.R
\name{normalize_df}
\alias{normalize_df}
\title{This function normalizes the n umeric columns in a data frame by
subtracting the mean and dividing by the standard deviation
given by the df from get_normalizing_factors}
\usage{
normalize_df(df, target_variable = NA, facs_df)
}
\arguments{
\item{df}{a data frame}

\item{target_variable}{a string name for the target variable}

\item{facs_df}{the factors to normalize by}
}
\value{
a data frame with numerical columns normalized
}
\description{
This function normalizes the n umeric columns in a data frame by
subtracting the mean and dividing by the standard deviation
given by the df from get_normalizing_factors
}
