% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_data.R
\name{prepare_training_set}
\alias{prepare_training_set}
\title{this function takes a data frame and the target variable and
returns a data structure with everything needed to train and
predict with the model.}
\usage{
prepare_training_set(df, target_variable = "y")
}
\arguments{
\item{df}{a data frame}

\item{target_variable}{a string name for the target variable}
}
\value{
list of data frames
}
\description{
this function takes a data frame and the target variable and
returns a data structure with everything needed to train and
predict with the model.
}
