% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_data.R
\name{rationalize_categoricals}
\alias{rationalize_categoricals}
\title{This function converts the categorical variables into either numeric (if ordered)
or characters, if not.}
\usage{
rationalize_categoricals(df, target_variable = "y")
}
\arguments{
\item{df}{a data frame}

\item{target_variable}{a string name for the target variable}
}
\value{
list
}
\description{
This function converts the categorical variables into either numeric (if ordered)
or characters, if not.
}
