% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_data.R
\name{remove_uninformative}
\alias{remove_uninformative}
\title{remove columns from the df that contain no information,
like factors with only one level}
\usage{
remove_uninformative(df)
}
\value{
data frame with uninformative columns removed
}
\description{
remove columns from the df that contain no information,
like factors with only one level
}
