% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{train_model}
\alias{train_model}
\title{this function returns a structure that contains the xgb model as well as
everything else it needs (normalization factors, levels of variables etc)
a predict function would need}
\usage{
train_model(train_structure, hyperparameters)
}
\arguments{
\item{train_structure}{the data structure generated by prepare_training_set}

\item{hyperparameters}{the list generated by guess_hyperparameters function}
}
\description{
this function returns a structure that contains the xgb model as well as
everything else it needs (normalization factors, levels of variables etc)
a predict function would need
}
