% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{train_rf_model}
\alias{train_rf_model}
\title{this function returns a structure that contains the rf model as well as
everything already contained in the train_xgb return structure}
\usage{
train_rf_model(train_structure, model_structure, hyperparameters)
}
\arguments{
\item{train_structure}{the data structure generated by prepare_training_set}

\item{model_structure}{the model structure created by train xgb}

\item{hyperparameters}{the list generated by guess_hyperparameters function}
}
\description{
this function returns a structure that contains the rf model as well as
everything already contained in the train_xgb return structure
}
