% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_data.R
\name{transform_target_variable}
\alias{transform_target_variable}
\title{This function converts the target variable to integer if it is a
categorical variable. and builds a map of the transformation.}
\usage{
transform_target_variable(df, target_variable)
}
\arguments{
\item{df}{a data frame}

\item{target_variable}{a string name for the target variable}
}
\value{
list
}
\description{
This function converts the target variable to integer if it is a
categorical variable. and builds a map of the transformation.
}
